+++ release/components/input/validators/max-value.directive.d.ts
@@ -1,10 +1,6 @@
-import { AfterViewInit, ElementRef } from '@angular/core';
 import { Validator, AbstractControl, ValidationErrors } from '@angular/forms';
-export declare class MaxValidatorDirective implements Validator, AfterViewInit {
-    private elm;
-    max: number;
-    type: string;
-    constructor(elm: ElementRef);
-    ngAfterViewInit(): void;
+export declare class MaxValidatorDirective implements Validator {
+    max: any;
+    type: any;
     validate(c: AbstractControl): ValidationErrors | null;
 }