+++ release/components/input/validators/min-value.directive.js
@@ -1,4 +1,4 @@
-import { Directive, Input, forwardRef, ElementRef } from '@angular/core';
+import { Directive, Input, forwardRef } from '@angular/core';
 import { NG_VALIDATORS, Validators } from '@angular/forms';
 var MIN_VALIDATOR = {
     provide: NG_VALIDATORS,
@@ -6,12 +6,8 @@ var MIN_VALIDATOR = {
     multi: true
 };
 var MinValidatorDirective = (function () {
-    function MinValidatorDirective(elm) {
-        this.elm = elm;
+    function MinValidatorDirective() {
     }
-    MinValidatorDirective.prototype.ngAfterViewInit = function () {
-        this.type = this.elm.nativeElement.getAttribute('type');
-    };
     MinValidatorDirective.prototype.validate = function (c) {
         if (this.type !== 'number') {
             return null;
@@ -21,15 +17,18 @@ var MinValidatorDirective = (function () {
     MinValidatorDirective.decorators = [
         { type: Directive, args: [{
                     selector: 'input[min]',
-                    providers: [MIN_VALIDATOR]
+                    providers: [MIN_VALIDATOR],
+                    host: {
+                        '[attr.min]': 'min ? min : null',
+                        '[attr.type]': 'type ? type : null'
+                    }
                 },] },
     ];
     /** @nocollapse */
-    MinValidatorDirective.ctorParameters = function () { return [
-        { type: ElementRef, },
-    ]; };
+    MinValidatorDirective.ctorParameters = function () { return []; };
     MinValidatorDirective.propDecorators = {
         'min': [{ type: Input },],
+        'type': [{ type: Input },],
     };
     return MinValidatorDirective;
 }());